Add ActivityRegistry
